backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
tortoise-orm[accel]==0.19.3
aerich==0.7.1
pydantic[email,dotenv]==1.10.7
python-multipart==0.0.6
